adjust tests accordingly
